refactor: 서버 시간 결과 ServerTimeResult 분리
